Added HashSet and TreeSet for FastUtil, JDK, and Trove

+++ scf4j-api/src/main/java/sg/yikjiun/scf4j/set/ByteSet.java
@@ -30,7 +30,7 @@ public interface ByteSet {
 
     boolean remove(byte b);
 
-    byte size();
+    int size();
 
     byte[] toArray();
 }
