+++ scf4j-api/src/main/java/sg/yikjiun/scf4j/set/DoubleSet.java
@@ -30,7 +30,7 @@ public interface DoubleSet {
 
     boolean remove(double d);
 
-    double size();
+    int size();
 
     double[] toArray();
 }
